refactor(std/io) Move LimitedReader to readers.ts (#6019)
